List of all functions

These are all functions that are available in the package, which can roughly be divided into two groups (running & reading LaMEM)

Running LaMEM

LaMEM.Run.remove_popup_messages_mac — Method

remove_popup_messages_mac()

On a Mac with firewall enabled, running LaMEM will result in a popup window that says: "Accept incoming connections" which you should Allow or Deny. This is a bit annoying, so this julia script fixes that. Note that you must have administrator rights on your machine as we need to run "sudo"

Run this script from the terminal with

julia> remove_popup_messages_mac()

You need to do this once (every time a new version is installed)

LaMEM.Run.run_lamem — Function

run_lamem(ParamFile::String, cores::Int64=1, args:String=""; wait=true)

This starts a LaMEM simulation, for using the parameter file ParamFile on cores number of cores. Optional additional command-line parameters can be specified with args.

Example:

You can call LaMEM with:

julia> using LaMEM
julia> ParamFile="../../input_models/BuildInSetups/FallingBlock_Multigrid.dat";
julia> run_lamem(ParamFile)

Do the same on 2 cores with a command-line argument as:

julia> ParamFile="../../input_models/BuildInSetups/FallingBlock_Multigrid.dat";
julia> run_lamem(ParamFile, 2, "-nstep_max = 1")

LaMEM.Run.run_lamem_save_grid — Function

ProcessorPartFile = run_lamem_save_grid(ParamFile::String, cores::Int64=1; verbose=true)

This calls LaMEM simulation, for using the parameter file ParamFile and creates processor partitioning file "ProcessorPartitioningcorescpuX.Y.Z.bin" for cores number of cores.

Example:

julia> using LaMEM
julia> ParamFile="../../input_models/BuildInSetups/FallingBlock_Multigrid.dat";
julia> ProcessorPartFile = run_lamem_save_grid(ParamFile, 2)

LaMEM.Run.show_paths_LaMEM — Method

show_paths_LaMEM()

The downloaded LaMEM binaries can also be called from outside julia (directly from the terminal). In that case, you will need to set load correct dynamic libraries (such as PETSc) and call the correct binaries.

This function shows this for your system.

Reading LaMEM output back into julia

```@autodocs Modules = [LaMEM.IO_functions]