List of all functions

These are all functions that are available in the package:

LaMEM.ReadField_3D_pVTR — Method

output, isCell = ReadField_3D_pVTR(data, FieldName::String)

Extracts a 3D data field from a pVTR data structure data

Input:

data: Data structure obtained with ReadVTRFile

FieldName: Exact name of the field as specified in the *.vtr file

Output:

data_field, isCell: 3D field with data, and a flag that indicates whether it is Cell data (PointData otherwise) data_field is a tuple of size 1, or 3 depending on whether it is a scalar or vector field

LaMEM.ReadField_3D_pVTU — Method

output, isCell = ReadField_3D_pVTU(data, FieldName::String)

Extracts a 3D data field from a pVTU data structure data

Input:

data: Data structure obtained with ReadVTRFile

FieldName: Exact name of the field as specified in the *.vtr file

Output:

data_field: Array with data, data_field is a tuple of size 1, 3 or 9 depending on whether it is a scalar, vector or tensor field

LaMEM.Read_VTR_File — Method

data_output = Read_VTR_File(DirName, FileName; field=nothing)

Reads a 3D LaMEM timestep from VTR file FileName, located in directory DirName. By default, it will read all fields. If you want you can only read a specific field. See the function fieldnames to get a list with all available fields in the file.

It will return data_output which is a CartData output structure.

LaMEM.Read_VTU_File — Method

data_output = Read_VTU_File(DirName, FileName; field=nothing)

Reads a 3D LaMEM timestep from VTU file FileName, located in directory DirName. Typically this is done to read passive tracers back into julia. By default, it will read all fields. If you want you can only read a specific field. See the function fieldnames to get a list with all available fields in the file.

It will return data_output which is a CartData output structure.

LaMEM.changefolder — Method

changefolder()

Starts a GUI on Windowss or Mac machines, which allows you to change our working directory

LaMEM.clean_directory — Function

clean_directory(DirName)

Removes all LaMEM timesteps & *.pvd files from the directory DirName

LaMEM.field_names — Method

names = field_names(DirName, FileName)

Returns the names of all fields stored in the vtr file in the directory Directory and file FileName.

LaMEM.field_names — Method

names = field_names(data)

Returns the names of all fields stored in the vtr data structure data.

LaMEM.readPVD — Method

FileNames, Time = readPVD(FileName::String)

This reads a PVD file & returns the timesteps and corresponding filenames

LaMEM.remove_popup_messages_mac — Method

remove_popup_messages_mac()

On a Mac with firewall enabled, running LaMEM will result in a popup window that says: "Accept incoming connections" which you should Allow or Deny. This is a bit annoying, so this julia script fixes that. Note that you must have administrator rights on your machine as we need to run "sudo"

Run this script from the terminal with

julia> remove_popup_messages_mac()

You need to do this once (every time a new version is installed)

LaMEM.run_lamem — Function

run_lamem(ParamFile::String, cores::Int64=1, args:String="")

This starts a LaMEM simulation, for using the parameter file ParamFile on cores number of cores. Optional additional command-line parameters can be specified with args.

Example:

The first step is to ensure that LaMEM_jll is installed on your system. You only need to do this once, or once LaMEM_jll is updated.

julia> import Pkg
julia> Pkg.add("LaMEM_jll")

Next you can call LaMEM with:

julia> ParamFile="../../input_models/BuildInSetups/FallingBlock_Multigrid.dat";
julia> run_lamem(ParamFile)

Do the same on 2 cores with

julia> ParamFile="../../input_models/BuildInSetups/FallingBlock_Multigrid.dat";
julia> run_lamem(ParamFile, 2, "-nstep_max = 1")

LaMEM.show_paths_LaMEM — Method

show_paths_LaMEM()

The downloaded LaMEM binaries can also be called from outside julia (directly from the terminal). In that case, you will need to set load correct dynamic libraries (such as PETSc) and call the correct binaries.

This function shows this for your system.